services/realtime-processing-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
httpx==0.25.0

# Kafka
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# WebSocket
websockets>=11.0
python-socketio>=5.10.0

# Async
asyncio
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

